deploy perdana ke vercel

+++ app/Http/Controllers/FakultasController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Fakultas;
 use Illuminate\Http\Request;
 
+
 class FakultasController extends Controller
 {
     /**
@@ -31,6 +32,11 @@ class FakultasController extends Controller
      */
     public function store(Request $request)
     {
+        // cek apakah user memiliki izin umtuk membuat fakultas
+        if($request->user()->cannot('create', Fakultas::class)) {
+            abort(403);
+        }
+
         // validasi input
         $input = $request->validate([
             'nama' => 'required|unique:fakultas',
@@ -74,6 +80,11 @@ class FakultasController extends Controller
     {
         $fakultas = Fakultas::findOrFail($fakultas);
 
+        // cek apakah user memiliki izin umtuk mengedit fakultas
+        if($request->user()->cannot('update', $fakultas)) {
+            abort(403);
+        }
+
         // validasi input (copas dari store)
         $input = $request->validate([
             'nama' => 'required',
@@ -92,10 +103,15 @@ class FakultasController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($fakultas)
+    public function destroy(Request $request, $fakultas)
     {
         $fakultas = Fakultas::findOrFail($fakultas);
         // dd($fakultas);
+
+        // cek apakah user memiliki izin umtuk menghapus fakultas
+        if($request->user()->cannot('destroy', $fakultas)) {
+            abort(403);
+        }
 
         // hapus data fakultas
         $fakultas->delete();
